+++ test-apps/pico-args-app/app.rs
@@ -24,8 +24,8 @@ fn main() {
     }
 }
 
-fn submain() -> Result<(), Box<std::error::Error>> {
-    let mut args = Arguments::from_env();
+fn submain() -> Result<(), pico_args::Error> {
+    let mut args = Arguments::from_env()?;
     let args = AppArgs {
         help: args.contains(["-h", "--help"]),
         number: args.value_from_str("--number")?.unwrap_or(5),